Using msgpack for passing control messages
